requirements.txt
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.10
cloudinary==1.29.0
Django==4.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
idna==3.3
pycodestyle==2.8.0
PyJWT==2.3.0
python-decouple==3.6
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
urllib3==1.26.8
